+++ lib/src/commands/sdk_commands/sdk_command.dart
@@ -1,4 +1,5 @@
 import 'package:args/command_runner.dart';
+import 'package:flutter_compile/src/commands/sdk_commands/_sdk_exec.dart';
 import 'package:flutter_compile/src/commands/sdk_commands/_sdk_global.dart';
 import 'package:flutter_compile/src/commands/sdk_commands/_sdk_install.dart';
 import 'package:flutter_compile/src/commands/sdk_commands/_sdk_list.dart';
@@ -14,6 +15,7 @@ import 'package:mason_logger/mason_logger.dart';
 /// `flutter_compile sdk remove <version>`
 /// `flutter_compile sdk global [version]`
 /// `flutter_compile sdk use [version]`
+/// `flutter_compile sdk exec <command> [args...]`
 ///
 /// A [Command] to manage multiple Flutter SDK versions.
 ///
@@ -26,6 +28,7 @@ class SdkCommand extends Command<int> {
     addSubcommand(SdkRemoveSubCommand(_logger));
     addSubcommand(SdkGlobalSubCommand(_logger));
     addSubcommand(SdkUseSubCommand(_logger));
+    addSubcommand(SdkExecSubCommand(_logger));
   }
 
   @override